refactor: 오류 사항 수정

+++ data/src/main/java/com/example/data/network/map/MapApi.kt
@@ -31,7 +31,7 @@ interface MapApi {
 
     @GET("/member-place/{memberPlaceId}")
     suspend fun getMemberPlaceDetail(
-        @Path("memberPlaceId") memberPlaceId: Long,
+        @Path("memberPlaceId") memberPlaceId: Int,
         @Query("size") size: Int? = 30,
         @Query("cursor") cursor: Long? = null
     ): Response<MemberPlaceDetailResponse>
